06: broken part 2 attempt
This code does not solve part 2, and I don't think it's on the right
path. The basic ideas of the attempt were

- focus on the *corners* in the path of the guard.
- each three consecutive corner defines a rectangle.
- if the missing corner can be reached from the 3rd
  corner w/out obstruction, it represents a place to form
  a cycle.
- if the mising corner *can't* be reached, find that that
  reachable point, and use replace the three corners
  with p2, p3, and the reachable point.

This finds 5 / 6 cases in the sample test data, but not all of
them. But looking at the missing case, the necessary obstacle is not
defined by any 3 consecutive corners in the path of the guard.

So, from peeking slightly at the spoilers from someone who got it,
it appears you have to watch for self-intersections while following
the path, and more generall (I think) for the opportunity to turn right
at a point that will intersect an earlier part of the path going in the
same direction.