06: part 2 solved
Finally. The key breakthough is not being so smart. The following are
the key realizations:

- You only need to consider places along the original path (without any
  extra obstacles in the grid) as locations for the obstacles.

- You have a cycle when the track revisits the same location with the
  same directon.

- Get the original path, and for each location on that path, place an
  obstacle there, compute the track in that grap, checking for a cycle
  as you go, remove the obstacle, and repeat for each location in the
  original path

- Again, generators for the win.

This solution is O(pathlen^2), which isn't great, but the input pathlen
is only a few hundred steps long.

Also, it's necessary to allow fo multiple right hand turns in a single
spot while generating the track.