07: use recursive backtracking
The previous code generated all possible operator placements, checking
for ones that evaluated to the desire result. This works but is O(2^n)
for part 1 and O(3^n) for part 2.

This commit reduces the the run time by usin a recursive backtracking
algorithm that works back from the *end* of the operands, checking if,
for each operator +, *, and ||, there is some value `new_target` such
that

    new_target op current_operand = current_target

and recursing on the new reduced targets. In practice this runs much
more quickly because impossible assignments can be pruned more quickly.