07: an iterative version of the backtracking solver
I got curious about the mechanics of converting a the recursive solver
(which is not tail-recursive) into an iterative form with an explicit
stack.

The hard problem is representing the suspended computation that is
normally captured by the normal call stack. But, generators to the
rescue. The "recursive" routine becomes a generator that yields actions
indicating what it's doing next:

- ("recur", args): start a new instance of the generator with the given
  arguments.

- ("return", value): this generator is done (ie, the "stack frame" is
  returning"), so deliver value to the parent / caller, which is next on
  the stack.

This was surprisingly easy to figure out, and even better it gets the
same answer.