aoc: introduce a place to collect handy utility functions
Up until now, I've just been copying puzzle .py files from one
day to the next, collecting a bunch of stuff that may or may not
be useful to any given day's puzzle, and probably more irritating taking
up a lot of space in the file.

Start the process of moving some of the utilities to a shared location.

- aoc/intertools.py: a dozen or so little iterator / generator / iterable
  functions that I've found useful