aoc/perf.py: timer class / context manager
This timer class / context manager was previously directly in the python
files for each puzzle, and just copied from one to another. It's time
for it to have it's own home, so here it is now.

The `timer` class provides a simple stop-watch like interface with three
public methods:

- `start()` starts the timer running,
- `stop()` stops the timer, and
- `elapsed()` meaured the elapsed time, which is either the time between
  starting and stopping the timer, if the timer is now stopped, or the
  time since the call to start.

The timer object can also be used as a context manager, eg,

```py
    with timer() as t:
        # do some stuff
    print(t.elapsed())
```

The timer context manager does not trap any exception, but it does stop
the timer on exit from the with block for any reason.