aoc.search: add `allpaths`
Add a function to enumerate all non-cyclic paths in a graph starting
from a given node. Calls look like

    allpaths([start], {start}, neighbours, keep)

where `start` is the initial node of the search, neighbours is a
function that takes a node and returns an iterable of neighbour nodes,
and keep is a predicate that controls which paths will be returned. If
keep is omitted, it defaults to True and all possible paths from start
are produced.

`allpaths` is a generator, so to collect all the paths:

    list(allpaths(...))