12: broken attempt at part 2
This solution solves the several sample inputs for part 2, *including*
the one where the edges you have to find are direction sensitive.

At the very beginning, I started down path of walking the perimeter,
which of course turns out to be the right thing, but I thought it would
be easier just to find horizontal and vertical sequences of edges, and
call those sides.

That works for all the sample inputs except ab.txt, because directional
perimeter intersects there, and that breaks a vertical column into *two*
edges, not one.

I tried to hack in a (not so) quick fix for that (the cross-edge stuff)
and that makes ab.txt work correctly, but the full input is still wrong.

Sigh...