13: solve part 2
Very easy to extend part 1 to part 2 with the linear algebra approach,
but I did have to reduce the tolerance to capture all the solutions that
were "close enough" to integers. (Apparently an epsilon of 1e-4 is ok
for AOC, but in my recollection, that would be a bit sloppy over in
actual math land.)

I'm not sure the diophantine equation approach would have worked at all,
although maybe. It would be critical to convert the possible solutions
into a point-slope form, and check for various intersections, rather
than trying to iteratively construct solutions derived from the initial
one.