18: performance improvements
brought the time for the full input down from 61s to 2.2s on my laptop.
the basic idea is to

- note that adding an obstacle can only make any path longer not
  shorter,

- which means that after adding an obstacle, the previous fastest path
  is still good *unless* it's now blocked.

So, save the shortest path after each a* search, and then for each block
placed, check if last saved path is still clear. If not, it's still the
shortest path, and more importantly the end position is still reachable,
so skip the astar step for this obstacle.