18: more part 2 performance improvements
bring the runtiem of part 2 down from 2.2 to about .78s, about a 3x
improvement, by *multiplying* the distance estimate by 250. This makes
the distance inadmissable, in a* terminology, so we may not be finding
the shortest path any more, but it does seem to make a big improvement
in finding *a* path.