16: a bit mess of broken part2 attempts
I'm still not fighting my way though part 2. After trying at least three
different implementations of finding all grid tiles on the shortest
path, it's always been the same number and that's always been wrong for
the test input but frustratingly *right* for the sample inputs.

I'm going to start over with a new idea for representing the search, and
this commit just captures the last broken state so that I can come back
here to poke around if I want to.